Fix positioning video elements using z-index
Instead of adding the video element view as a su-view of the WebView,
add it as a subview of it's superview. This puts the WebView and the
video views at the same level (aka, they are siblings) and the layer's
zPosition is considered amongst siblings.

Since an ASCII diagram is worth a thousand images, this is the layout
before this patch:

+-------------------------------+
|           superview           |
+-------------------------------+
+-------------------------------+
|            WebView            |
+-------------------------------+
+--------------+ +--------------+
|  videoView1  | |  videoView2  |
+--------------+ +--------------+

And after:

+---------------------------------------+
|                superview              |
+---------------------------------------+
+---------+ +------------+ +------------+
| WebView | | videoView1 | | videoView2 |
+---------+ +------------+ +------------+

This makes it possible to order videos "behind" the WebView, which this
patch makes transparent, and thus have HTML elements on top of the video
views.

Props to @1N50MN14 for writing a first version of (what became) this patch.