Uncomment, and fix, onGetStatsCallback closure
While working on a project using WebRTC with my friend @peili, we unearthed a bug where the getStats() callback closure was not being called. We noticed it was commented out as part of the Swift 3 syntax migration, and never added back. This correctly uses the expected parameter for the closure (an NSArray), which results in a correct callback, and result, of getStats().